# server/openapi.yaml
openapi: 3.1.0
info:
  title: Lumina Backend API
  version: 1.0.0
  description: |
    주차장 사고 증거 매칭 데모 MVP 백엔드 (BE_SPEC.md F0~F10).

    공통 규약
    - 모든 응답은 `{ data, meta }` 또는 `{ error: { code, message, retryable, details? }, meta: { requestId } }`.
    - 인증: `Authorization: Bearer <Supabase access_token>`. `/config`, `/health`, `/ready`, `/demo/*` 만 공개.
    - 시간은 ISO 8601 UTC 문자열. 금액은 KRW 정수.
    - 보험사 채택·정산은 Mock 이며 관련 DTO 는 항상 `mock: true` 를 포함한다.
    - AI 결과(`analysis.result`)는 사고 사실·가해 차량·과실을 확정하지 않는다 (`disclaimer` 참고).
servers:
  - url: http://localhost:3001/api/v1
    description: local
...
security:
  - bearerAuth: []
paths:
...
  /uploads/photo-url:
    post:
      tags: [F3 incidents]
      summary: 피해 사진 staging 업로드용 signed URL (REQUESTER)
      description: |
        `staging/{userId}/{uuid}.{ext}` 경로로 signed upload URL 을 발급한다. 업로드 후 `POST /incidents` 의
        `photoObjectPaths` 에 넘기면 사고 경로로 이동·메타데이터 제거 후 연결된다. 24h(STAGING_RETENTION_HOURS) 내
        연결되지 않은 객체는 자동 삭제. Supabase signed upload URL 은 2시간 고정이므로 `expiresAt` 은 스토리지 값이다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mime, bytes]
              additionalProperties: false
              properties:
                mime: { type: string, enum: [image/jpeg, image/png, image/webp] }
                bytes: { type: integer, minimum: 1 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/SignedUpload' }
        '403': { $ref: '#/components/responses/ForbiddenRole' }
        '413': { description: FILE_TOO_LARGE }
        '415': { description: UNSUPPORTED_MEDIA_TYPE }
  /incidents:
    post:
      tags: [F3 incidents]
      summary: 사고 요청 등록 (REQUESTER) → 예치(Mock) 기록 → F4 매칭·알림
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncidentCreate' }
      responses:
        '201':
          description: 'Created. 매칭 결과 포함. 커밋 후 DTO 생성이 실패하면 `{ id, status, matching, partial: true }` 만 반환.'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/IncidentOwner'
                          - properties:
                              matching: { $ref: '#/components/schemas/MatchingResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/ForbiddenRole' }
        '415': { description: staging 사진 MIME/시그니처 불일치 }
  /incidents/map:
    get:
      tags: [F3 incidents]
      summary: 로그인한 모든 사용자의 지도용 신고 목록
      description: 공개된 신고의 장소와 좌표만 반환. 최대 200건이며 hasMore가 true이면 지도를 확대한다. 범위는 네 값을 모두 지정하거나 모두 생략한다.
      parameters:
        - { name: south, in: query, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: north, in: query, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: west, in: query, schema: { type: number, minimum: -180, maximum: 180 } }
        - { name: east, in: query, schema: { type: number, minimum: -180, maximum: 180 } }
      responses:
        '200':
          description: 지도 영역에 있는 신고
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          hasMore: { type: boolean }
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                id: { type: string, format: uuid }
                                place: { $ref: '#/components/schemas/Place' }
                                type: { $ref: '#/components/schemas/IncidentType' }
                                status: { $ref: '#/components/schemas/IncidentStatus' }
                                occurredFrom: { type: string, format: date-time }
                                occurredTo: { type: string, format: date-time }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /incidents/{id}:
    get:
      tags: [F3 incidents]
      summary: 사고 요청 상세. 소유자/OPERATOR 는 전체, 로그인한 나머지 사용자는 마스킹 DTO
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/IncidentOwner'
                          - $ref: '#/components/schemas/IncidentWitnessView'
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Supabase access token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    Unauthenticated:
      description: UNAUTHENTICATED
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ForbiddenRole:
      description: FORBIDDEN_ROLE
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: NOT_FOUND (접근 권한 없는 리소스도 404)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ValidationError:
      description: VALIDATION_ERROR (details 에 zod issues)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          type: object
          properties:
            requestId: { type: string }
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            retryable: { type: boolean }
            details: {}
        meta:
          type: object
          properties:
            requestId: { type: string }
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR      # 400
        - UNAUTHENTICATED       # 401
        - FORBIDDEN_ROLE        # 403
        - NOT_FOUND             # 404
        - INVALID_STATE         # 409
        - ALREADY_EXISTS        # 409
        - FILE_TOO_LARGE        # 413
        - UNSUPPORTED_MEDIA_TYPE # 415
        - VIDEO_UNANALYZABLE    # 422 (too short/long, unreadable MP4 header)
        - RATE_LIMITED          # 429
        - INTERNAL_ERROR        # 500
        - PROVIDER_UNAVAILABLE  # 503
...
    Place:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: A주차장 }
        kind: { type: string, enum: [PARKING_LOT, APARTMENT, BUILDING] }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
    SignedUpload:
      type: object
      properties:
        objectPath: { type: string }
        uploadUrl: { type: string }
        token: { type: string }
        expiresAt: { type: string, format: date-time, description: 스토리지가 실제로 강제하는 만료 시각 }
    IncidentType: { type: string, enum: [HIT_AND_RUN, CONTACT, DAMAGE, OTHER] }
    IncidentStatus: { type: string, enum: [DRAFT, OPEN, COLLECTING, REVIEWING, ADOPTED, CLOSED_NO_EVIDENCE, CANCELLED] }
    Vehicle:
      type: object
      required: [color, model, damageArea]
      properties:
        color: { type: string, maxLength: 50 }
        model: { type: string, maxLength: 100 }
        damageArea: { type: string, maxLength: 100 }
    IncidentCreate:
      type: object
      required: [type, occurredFrom, occurredTo, vehicle, description, consent]
      anyOf:
        - required: [placeId]
        - required: [location]
      additionalProperties: false
      properties:
        placeId: { type: string, format: uuid }
        location:
          type: object
          additionalProperties: false
          required: [name, address, lat, lng]
          properties:
            name: { type: string, minLength: 1, maxLength: 200 }
            address: { type: string, minLength: 1, maxLength: 500 }
            lat: { type: number, minimum: -90, maximum: 90 }
            lng: { type: number, minimum: -180, maximum: 180 }
        type: { $ref: '#/components/schemas/IncidentType' }
        occurredFrom: { type: string, format: date-time }
        occurredTo: { type: string, format: date-time, description: occurredFrom 이후, 최대 24h 창 }
        vehicle: { $ref: '#/components/schemas/Vehicle' }
        description: { type: string, maxLength: 1000 }
        photoObjectPaths:
          type: array
          maxItems: 10
          items: { type: string }
          description: '`POST /uploads/photo-url` 로 받은 staging 경로 (본인 것만)'
        consent:
          type: object
          required: [evidenceUse, privacy]
          properties:
            evidenceUse: { type: boolean, const: true }
            privacy: { type: boolean, const: true }
    Photo:
      type: object
      properties:
        id: { type: string, format: uuid }
        position: { type: integer }
        mime: { type: string }
        width: { type: integer }
        height: { type: integer }
        url: { type: [string, 'null'], description: signed URL (스토리지 장애 시 null) }
        expiresAt: { type: [string, 'null'], format: date-time }
    MatchingResult:
      type: object
      properties:
        matchedWitnessCount: { type: integer }
        notifiedAt: { type: [string, 'null'], format: date-time }
    IncidentOwner:
      type: object
      description: 소유자(X)/OPERATOR 용
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        reviewMode: { type: string, example: AUTO_DEMO }
        type: { $ref: '#/components/schemas/IncidentType' }
        place: { $ref: '#/components/schemas/Place' }
        occurredFrom: { type: string, format: date-time }
        occurredTo: { type: string, format: date-time }
        vehicle: { $ref: '#/components/schemas/Vehicle' }
        description: { type: string }
        photos: { type: array, items: { $ref: '#/components/schemas/Photo' } }
        matchedWitnessCount: { type: integer }
        submissions:
          type: object
          properties:
            total: { type: integer }
            ready: { type: integer }
        settlement:
          oneOf: [{ $ref: '#/components/schemas/Settlement' }, { type: 'null' }]
        publishedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    IncidentWitnessView:
      type: object
      description: 로그인한 일반 조회자 용 마스킹 DTO (요청자 신원·연락처 없음)
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        type: { $ref: '#/components/schemas/IncidentType' }
        place: { $ref: '#/components/schemas/Place' }
        occurredFrom: { type: string, format: date-time }
        occurredTo: { type: string, format: date-time }
        vehicle: { $ref: '#/components/schemas/Vehicle' }
        descriptionSummary: { type: string, maxLength: 201 }
        photos:
          type: array
          items:
            type: object
            properties:
              position: { type: integer }
              url: { type: [string, 'null'] }
              width: { type: integer }
              height: { type: integer }
        rewardPreview:
          type: object
          properties:
            amount: { type: integer, example: 80000 }
            mock: { type: boolean, const: true }
        mySubmissionId: { type: [string, 'null'], format: uuid }
        masked: { type: boolean, const: true }
        createdAt: { type: string, format: date-time }
...
    SettlementStatus:
      type: string
      enum: [DEPOSIT_PENDING, DEPOSITED, ADOPTION_PENDING, PAYOUT_SCHEDULED, PAID, REFUNDED, PAYOUT_FAILED, DISPUTED]
    Settlement:
      type: object
      properties:
        id: { type: string, format: uuid }
        incidentId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/SettlementStatus' }
        depositAmount: { type: integer, example: 100000 }
        platformFee: { type: integer, example: 20000 }
        witnessReward: { type: integer, example: 80000 }
        payoutScheduledAt: { type: [string, 'null'], format: date-time }
        mock: { type: boolean, const: true }
        label: { type: string, const: '데모: 실제 결제·송금 없음' }
...